Safari-611.1.10/LayoutTests/editing/inserting/insert-list-in-table-cell-03.html

<!DOCTYPE html>
<div id="container" contenteditable="true"><table border="1"><tbody id="element"><tr><td>fsdf</td><td>fsdf</td></tr><tr><td>gghfg</td><td>fsfg</td></tr></tbody></table></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
    Markup.description('Insert list items in all the cells of a table:');

    var e = document.getElementById("element");
    window.getSelection().selectAllChildren(e);
    Markup.dump('container', 'Before');

    document.execCommand("insertOrderedList");
    Markup.dump('container', 'After');
</script>
